Fix EL tests.
