+++ spacy_llm/tasks/entity_linker/candidate_selector.py
@@ -1,0 +1,83 @@
+import csv
+import warnings
+from pathlib import Path
+from typing import Dict, Iterable, Union
+
+import spacy
+from spacy.pipeline import EntityLinker
+from spacy.tokens import Span
+
+from .ty import UNAVAILABLE_ENTITY_DESC, Entity
+
+
+class PipelineCandidateSelector:
+    """Callable generated by loading and wrapping a spaCy pipeline with an NEL component and a filled knowledge base."""
+
+    def __init__(
+        self, nlp_path: Union[Path, str], desc_path: Union[Path, str], top_n: int = 5
+    ):
+        """
+        Loads spaCy pipeline, knowledge base, entity descriptions.
+        top_n (int): Top n candidates to include in prompt.
+        nlp_path (Union[Path, str]): Path to stored spaCy pipeline.
+        desc_path (Union[Path, str]): Path to .csv file with descriptions for entities. Has to have two columns
+          with the first one being the entity ID, the second one being the description. The entity ID has to match with
+          the entity ID in the stored knowledge base.
+        """
+        self._nlp = spacy.load(nlp_path)
+        if "entity_linker" not in self._nlp.component_names:
+            raise ValueError(
+                f"'entity_linker' component has to be available in specified pipeline at {nlp_path}, but "
+                f"isn't."
+            )
+        self._entity_linker: EntityLinker = self._nlp.get_pipe("entity_linker")
+        self._kb = self._entity_linker.kb
+        with open(desc_path) as csvfile:
+            self._descs: Dict[str, str] = {}
+            for row in csv.reader(csvfile, quoting=csv.QUOTE_ALL, delimiter=";"):
+                if len(row) != 2:
+                    continue
+                self._descs[row[0]] = row[1]
+
+            if len(self._descs) == 0:
+                raise ValueError(
+                    "Format of CSV file with entity descriptions is wrong. CSV has to be formatted as "
+                    "semicolon-delimited CSV with two columns. The first columns has to contain the entity"
+                    " ID, the second the entity description."
+                )
+        self._top_n = top_n
+
+    def __call__(self, mentions: Iterable[Span]) -> Iterable[Iterable[Entity]]:
+        """Retrieves top n candidates using spaCy's entity linker's .get_candidates_batch().
+        mentions (Iterable[Span]): Mentions to look up entity candidates for.
+        RETURNS (Iterable[Iterable[Entity]]): Top n entity candidates per mention.
+        """
+        all_cands = self._kb.get_candidates_batch(mentions)
+        for cands in all_cands:
+            assert isinstance(cands, list)
+            cands.sort(key=lambda x: x.prior_prob, reverse=True)
+
+        return [
+            [
+                Entity(
+                    id=cand.entity_,
+                    description=self.get_entity_description(cand.entity_),
+                )
+                for cand in cands[: self._top_n]
+            ]
+            for cands in all_cands
+        ]
+
+    def get_entity_description(self, entity_id: str) -> str:
+        """Returns entity description for entity ID. If none found, a warning is emitted and
+            spacy_llm.tasks.enttiy_linker.ty.UNAVAILABLE_ENTITY_DESC is returned.
+        entity_id (str): Entity whose ID should be looked up.
+        RETURNS (str): Entity description for entity with specfied ID. If no description found, returned string equals
+            spacy_llm.tasks.enttiy_linker.ty.UNAVAILABLE_ENTITY_DESC.
+        """
+        if entity_id not in self._descs:
+            warnings.warn(
+                f"Entity with ID {entity_id} is not in provided descriptions file."
+            )
+
+        return self._descs.get(entity_id, UNAVAILABLE_ENTITY_DESC)
