+++ spacy_llm/tests/tasks/examples/entity_linker.yml
@@ -55,4 +55,4 @@
   reasons:
     - The mention of "Big Apple" in the same context clarifies that this is about the city New York.
     - Big Apple is a well-known nickname of New York.
-    - The context of "stores" indicates that this is about the technology company Apple, who operates "Apple stores".
+    - The context of "stores" indicates that this is about the technology company Apple, which operates "Apple stores".
